+++ Team_03/02_Source/01_Source_Code/FE/mobile/lib/features/fitness/view/diary/diary_screen.dart
@@ -234,7 +234,7 @@ class _DiaryScreenState extends State<DiaryScreen>
     BuildContext context,
     String mealTitle,
     List<Food> foodItems,
-    int calories,
+      double calories,
     MealType mealType,
   ) {
     final textTheme = Theme.of(context).textTheme;
